+++ R/git.R
@@ -98,10 +98,15 @@ get_outdated_files <- function(repo, files, outdir = NULL) {
   # For each source file, determine if it has been committed more recently than
   # its corresponding HTML
   out_of_date <- logical(length = length(files))
+
+  blobs <- git2r::odb_blobs(repo)
+  blobs$fname <- ifelse(blobs$path == "", blobs$name,
+                        file.path(blobs$path, blobs$name))
+
   for (i in seq_along(files)) {
     # Most recent commit time of source and HTML files
-    recent_source <- get_recent_commit_time(repo, files[i])
-    recent_html <- get_recent_commit_time(repo, html[i])
+    recent_source <- max(blobs$when[blobs$fname == files[i]])
+    recent_html <- max(blobs$when[blobs$fname == html[i]])
     if (recent_source >= recent_html) {
       out_of_date[i] <- TRUE
     }
@@ -110,25 +115,6 @@ get_outdated_files <- function(repo, files, outdir = NULL) {
   # Prepend path to Git repository
   outdated <- paste0(path_to_git, outdated)
   return(outdated)
-}
-
-# Get the time of the most recent commit for a file.
-#
-# repo: git_repository object
-# f: path to file relative to .git
-#
-# Note: This function is not vectorized.
-get_recent_commit_time <- function(repo, f) {
-  # Obtain every commit for the file
-  blame <- git2r::blame(repo, f)
-  # Extract the times of the commits
-  times <- sapply(blame@hunks,
-                  function(x) git2r::when(x@final_signature@when))
-  times <- strptime(times, format = "%Y-%m-%d %H:%M:%S")
-  times <- sort(unique(times), decreasing = TRUE)
-  # Most recent commit time
-  recent <- times[1]
-  return(recent)
 }
 
 # Decide which files to render and commit